% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalisation.R
\name{coloc}
\alias{coloc}
\title{Compute a colocalisation}
\usage{
coloc(self, ...)
}
\arguments{
\item{self}{a colocRedRibbon object}
}
\value{
RedRibbonColoc object
}
\description{
Compute a colocalisation
}
