% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalisation.R
\name{RedRibbonColoc}
\alias{RedRibbonColoc}
\title{Compute an enrichment based colocalization}
\usage{
RedRibbonColoc(
  data,
  algorithm = c("ea", "classic"),
  half = 6300,
  niter = 96,
  risk = NULL,
  effect = `>=`,
  columns = NULL,
  shortlist = TRUE
)
}
\arguments{
\item{data}{a data.frame with columns id, a, b, position. For respectively, the name
of the SNP, the p-value of the first analysis, the p-value of the second analysis, and
the position on the chromosome. Additionnal columns are for colocalisation:
\link{ab}.n (number of samples of the analysis), \link{ab}.eaf (effect allele frequency),
\link{ab}.or (odd-ratio) and/or \link{ab}.beta (effect slope).}

\item{algorithm}{the algorithm to use for minimal hypergeometric p-value searching}

\item{half}{the linkage desiquilibrium fitting function parameter for permutation}

\item{niter}{the number of iteration for adjusted p-value computation}

\item{risk}{the GWAS dataset with an odd-ratio (e.g. a.or or b.or) either NULL, 'a' or 'b'}

\item{effect}{an operator like \code{>=} or \code{<=} indicating the effect direction for the non-risk
dataset}

\item{columns}{a named vector with column names in the `data' data frame}

\item{shortlist}{run RedRibbon before coloc (default = TRUE)}
}
\value{
RedRibbonColoc object
}
\description{
Compute an enrichment based colocalization
}
